apps/higgs-tts/requirements.txt
transformers>=4.45.1,<4.47.0
librosa
soundfile
dacite
boto3==1.35.36
s3fs
json_repair
pandas
pydantic
vector_quantize_pytorch
loguru
pydub
ruff==0.12.2
omegaconf
click
langid
jieba
accelerate>=0.26.0
pyloudnorm
flatten_dict
julius
ffmpy
importlib_resources
randomname
argbind
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-logging
grpcio
grpcio-tools
grpcio-reflection
protobuf